# Fix: External sort failing on `StringView` due to shared buffers