@tustvold 2025-02-24
FWIW I think this compaction logic probably makes sense as part of the IPCWriter, as opposed to a workaround here - https://github.com/apache/arrow-rs/issues/7185